LayoutTests/fast/scrolling/scroll-anchoring/heuristic-enabled-below-threshold-expected.txt


Testing anchoring with 9 of size 10 - expect anchoring true
PASS anchoringActive is true
PASS successfullyParsed is true

TEST COMPLETE
